Moving exceptions to own files
